+++ src/Microsoft.ApplicationInsights/DataContracts/PageViewTelemetry.cs
@@ -18,11 +18,10 @@
     /// method.
     /// <a href="https://go.microsoft.com/fwlink/?linkid=525722#page-views">Learn more</a>
     /// </remarks>
-    public sealed class PageViewTelemetry : ITelemetry, ISupportProperties, ISupportSampling, ISupportMetrics
+    public sealed class PageViewTelemetry : ITelemetry, ISupportProperties, ISupportSampling, ISupportMetrics, IAiSerializableTelemetry
     {
         internal const string TelemetryName = "PageView";
 
-        internal const string BaseType = nameof(PageViewData);
         internal readonly PageViewData Data;
         private readonly TelemetryContext context;
         private IExtension extension;
@@ -59,6 +58,12 @@
             this.extension = source.extension?.DeepClone();
             this.Timestamp = source.Timestamp;
         }
+
+        /// <inheritdoc />
+        string IAiSerializableTelemetry.TelemetryName => TelemetryName;
+
+        /// <inheritdoc />
+        string IAiSerializableTelemetry.BaseType => nameof(PageViewData);
 
         /// <summary>
         /// Gets or sets date and time when event was recorded.