+++ src/Microsoft.ApplicationInsights/Extensibility/Implementation/JsonSerializer.cs
@@ -156,103 +156,14 @@
             return new GZipStream(stream, CompressionMode.Compress);
         }
 
-        /// <summary>
-        /// Copies GlobalProperties to the target, avoiding accessing the public accessor GlobalProperties
-        /// unless needed, to avoid the penality of ConcurrentDictionary instantiation. 
-        /// </summary>        
-        private static void CopyGlobalPropertiesIfExist(TelemetryContext context, IDictionary<string, string> target)
-        {
-            if (context.GlobalPropertiesValue != null)
-            {
-                Utils.CopyDictionary(context.GlobalProperties, target);
-            }
-        }
-
         private static void SerializeTelemetryItem(ITelemetry telemetryItem, JsonSerializationWriter jsonSerializationWriter)
         {
             jsonSerializationWriter.WriteStartObject();
 
-            if (telemetryItem is EventTelemetry)
+            if (telemetryItem is IAiSerializableTelemetry serializeableTelemetry)
             {
-                EventTelemetry eventTelemetry = telemetryItem as EventTelemetry;
-                CopyGlobalPropertiesIfExist(telemetryItem.Context, eventTelemetry.Data.properties);
-
-                SerializeHelper(telemetryItem, jsonSerializationWriter, EventTelemetry.BaseType, EventTelemetry.TelemetryName);
-            }
-            else if (telemetryItem is ExceptionTelemetry)
-            {
-                ExceptionTelemetry exTelemetry = telemetryItem as ExceptionTelemetry;
-                CopyGlobalPropertiesIfExist(telemetryItem.Context, exTelemetry.Data.Data.properties);
-
-                SerializeHelper(telemetryItem, jsonSerializationWriter, ExceptionTelemetry.BaseType, ExceptionTelemetry.TelemetryName);
-            }
-            else if (telemetryItem is MetricTelemetry)
-            {
-                MetricTelemetry mTelemetry = telemetryItem as MetricTelemetry;
-                CopyGlobalPropertiesIfExist(telemetryItem.Context, mTelemetry.Data.properties);
-
-                SerializeHelper(telemetryItem, jsonSerializationWriter, MetricTelemetry.BaseType, MetricTelemetry.TelemetryName);
-            }
-            else if (telemetryItem is PageViewTelemetry)
-            {
-                PageViewTelemetry pvTelemetry = telemetryItem as PageViewTelemetry;
-                CopyGlobalPropertiesIfExist(telemetryItem.Context, pvTelemetry.Data.properties);
-
-                SerializeHelper(telemetryItem, jsonSerializationWriter, PageViewTelemetry.BaseType, PageViewTelemetry.TelemetryName);
-            }
-            else if (telemetryItem is PageViewPerformanceTelemetry)
-            {
-                PageViewPerformanceTelemetry pvptelemetry = telemetryItem as PageViewPerformanceTelemetry;
-                CopyGlobalPropertiesIfExist(telemetryItem.Context, pvptelemetry.Data.properties);
-
-                SerializeHelper(telemetryItem, jsonSerializationWriter, PageViewPerformanceTelemetry.BaseType, PageViewPerformanceTelemetry.TelemetryName);
-            }
-            else if (telemetryItem is DependencyTelemetry)
-            {
-                DependencyTelemetry depTelemetry = telemetryItem as DependencyTelemetry;
-                CopyGlobalPropertiesIfExist(telemetryItem.Context, depTelemetry.InternalData.properties);
-
-                SerializeHelper(telemetryItem, jsonSerializationWriter, DependencyTelemetry.BaseType, DependencyTelemetry.TelemetryName);
-            }
-            else if (telemetryItem is RequestTelemetry)
-            {
-                RequestTelemetry reqTelemetry = telemetryItem as RequestTelemetry;
-                if (telemetryItem.Context.GlobalPropertiesValue != null)
-                {
-                    Utils.CopyDictionary(telemetryItem.Context.GlobalProperties, reqTelemetry.Properties);
-                }
-
-                // CopyGlobalPropertiesIfExist(telemetryItem.Context, reqTelemetry.Data.properties);
-
-                SerializeHelper(telemetryItem, jsonSerializationWriter, RequestTelemetry.BaseType, RequestTelemetry.TelemetryName);
-            }
-#pragma warning disable 618
-            else if (telemetryItem is PerformanceCounterTelemetry)
-            {
-                PerformanceCounterTelemetry pcTelemetry = telemetryItem as PerformanceCounterTelemetry;
-                CopyGlobalPropertiesIfExist(telemetryItem.Context, pcTelemetry.Data.Properties);
-
-                SerializeHelper(telemetryItem, jsonSerializationWriter, MetricTelemetry.BaseType, MetricTelemetry.TelemetryName);
-            }
-            else if (telemetryItem is SessionStateTelemetry)
-            {
-                SessionStateTelemetry ssTelemetry = telemetryItem as SessionStateTelemetry;
-                SerializeHelper(telemetryItem, jsonSerializationWriter, EventTelemetry.BaseType, EventTelemetry.TelemetryName);
-            }
-#pragma warning restore 618
-            else if (telemetryItem is TraceTelemetry)
-            {
-                TraceTelemetry traceTelemetry = telemetryItem as TraceTelemetry;
-                CopyGlobalPropertiesIfExist(telemetryItem.Context, traceTelemetry.Data.properties);
-
-                SerializeHelper(telemetryItem, jsonSerializationWriter, TraceTelemetry.BaseType, TraceTelemetry.TelemetryName);
-            }
-            else if (telemetryItem is AvailabilityTelemetry)
-            {
-                AvailabilityTelemetry availabilityTelemetry = telemetryItem as AvailabilityTelemetry;
-                CopyGlobalPropertiesIfExist(telemetryItem.Context, availabilityTelemetry.Data.properties);
-
-                SerializeHelper(telemetryItem, jsonSerializationWriter, AvailabilityTelemetry.BaseType, AvailabilityTelemetry.TelemetryName);
+                telemetryItem.CopyGlobalPropertiesIfExist();
+                SerializeHelper(telemetryItem, jsonSerializationWriter, telemetryName: serializeableTelemetry.TelemetryName, baseType: serializeableTelemetry.BaseType);
             }
             else
             {
@@ -279,7 +190,7 @@
         {
             DictionarySerializationWriter dictionarySerializationWriter = new DictionarySerializationWriter();
             telemetryItem.SerializeData(dictionarySerializationWriter); // Properties and Measurements are covered as part of Data if present
-            CopyGlobalPropertiesIfExist(telemetryItem.Context, dictionarySerializationWriter.AccumulatedDictionary);
+            telemetryItem.CopyGlobalPropertiesIfExist(dictionarySerializationWriter.AccumulatedDictionary);
 
             jsonSerializationWriter.WriteProperty("name", telemetryItem.WriteTelemetryName(EventTelemetry.TelemetryName));
             telemetryItem.WriteEnvelopeProperties(jsonSerializationWriter); // No need to copy Context - it's serialized here from the original item